requirements.txt
mcp>=0.1.0
msal>=1.20.0
requests>=2.28.0
pandas>=1.5.0
python-dotenv>=0.21.0